Update Dockerfile
Updated Dockerfile to include new dependencies as listed in fw-Plantiga_Osensa/.config/config.yml.

Added package installs that can be now removed from config.yml:
`default-jdk    zip    gcc-7`

Left PlantUML setup to be done in CircleCI commands to keep image clean, but it is commented in if we want to pack that into the image at some point.